pyproject.toml
[tool.poetry]
name = "cyclops"
version = "0.1.0"
description = ""
authors = ["Vivaan Verma"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dlib = "^19.24.5"
face-recognition = "^1.3.0"
numpy = "^2.0.1"
pillow = "^10.4.0"
face-recognition-models = {git = "https://github.com/ageitgey/face_recognition_models"}
setuptools = "^72.1.0"
typer = "^0.12.3"
rich = "^13.7.1"
opencv-python = "^4.10.0.84"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
